requirements.txt
requests>=2.31.0
openai>=1.3.0
python-dotenv>=1.0.0
rich>=13.0.0
click>=8.0.0
fastapi>=0.110.0
uvicorn>=0.23.0
feedparser>=6.0.10
